+++ yux/ast/node/fn_node.cpp
@@ -36,7 +36,6 @@ TypeInfo FnHeaderNode::getType() const {
 
 string FnHeaderNode::resolvedFallibleErr() const {
     if (_fallibleErrType) return _fallibleErrType->getType().name;
-    if (auto e = getAnnoArg("Fallible")) return *e;
     return {};
 }